Example that works with MySQL
This "sort" by price then name requires two subselects. The first finds
the min price for each category, and the second finds the min name.

If we were concerned about multiple records having the smame price and
name, a third subselect would be required to find the min id to throw
out the pseudo-duplicates